Refactoring integration tests
