+++ tests/fixes/duplicated_key.rs
@@ -1,14 +1,14 @@
-use crate::common::TestDir;
+use crate::common::*;
 
 #[test]
 fn duplicated_key() {
     let testdir = TestDir::new();
     let testfile = testdir.create_testfile(".env", "ABC=DEF\nABC=XYZ\nFOO=BAR\nFOO=BAZ\n");
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:2 DuplicatedKey: The ABC key is duplicated\n\
-        .env:4 DuplicatedKey: The FOO key is duplicated\n",
-    );
+    let expected_output = fix_output(&[
+        ".env:2 DuplicatedKey: The ABC key is duplicated",
+        ".env:4 DuplicatedKey: The FOO key is duplicated",
+    ]);
+
     testdir.test_command_fix_success(expected_output);
 
     assert_eq!(
