+++ tests/fixes/leading_character.rs
@@ -1,15 +1,14 @@
-use crate::common::TestDir;
+use crate::common::*;
 
 #[test]
 fn leading_character() {
     let testdir = TestDir::new();
     let testfile = testdir.create_testfile(".env", "*BAR=BAZ\n.ABC=DEF\n1QUX=QUUX\n_FOO=BAR\n");
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:1 LeadingCharacter: Invalid leading character detected\n\
-        .env:2 LeadingCharacter: Invalid leading character detected\n\
-        .env:3 LeadingCharacter: Invalid leading character detected\n",
-    );
+    let expected_output = fix_output(&[
+        ".env:1 LeadingCharacter: Invalid leading character detected",
+        ".env:2 LeadingCharacter: Invalid leading character detected",
+        ".env:3 LeadingCharacter: Invalid leading character detected",
+    ]);
 
     testdir.test_command_fix_success(expected_output);
