+++ tests/fixes/mod.rs
@@ -1,6 +1,6 @@
 mod quote_character;
 
-use crate::common::TestDir;
+use crate::common::*;
 
 mod duplicated_key;
 mod ending_blank_line;
@@ -30,11 +30,10 @@ fn skip_checks() {
     let testdir = TestDir::new();
     let testfile = testdir.create_testfile(".env", "A1=1\nA2=2\na0=0\na2=2\n");
 
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:3 LowercaseKey: The a0 key should be in uppercase\n\
-        .env:4 LowercaseKey: The a2 key should be in uppercase\n",
-    );
+    let expected_output = fix_output(&[
+        ".env:3 LowercaseKey: The a0 key should be in uppercase",
+        ".env:4 LowercaseKey: The a2 key should be in uppercase",
+    ]);
 
     testdir.test_command_fix_success_with_args(
         expected_output,
@@ -54,14 +53,13 @@ fn multiple_files() {
     let testfile2 = testdir.create_testfile("2.env", "abc=DEF\n\nF=BAR\nB=bbb\n");
     let testfile3 = testdir.create_testfile("3.env", "A=b \nab=DEF\n\nA=c\n");
 
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        2.env:1 LowercaseKey: The abc key should be in uppercase\n\
-        2.env:4 UnorderedKey: The B key should go before the F key\n\
-        3.env:1 TrailingWhitespace: Trailing whitespace detected\n\
-        3.env:2 LowercaseKey: The ab key should be in uppercase\n\
-        3.env:4 DuplicatedKey: The A key is duplicated\n",
-    );
+    let expected_output = fix_output(&[
+        "2.env:1 LowercaseKey: The abc key should be in uppercase",
+        "2.env:4 UnorderedKey: The B key should go before the F key",
+        "3.env:1 TrailingWhitespace: Trailing whitespace detected",
+        "3.env:2 LowercaseKey: The ab key should be in uppercase",
+        "3.env:4 DuplicatedKey: The A key is duplicated",
+    ]);
     testdir.test_command_fix_success(expected_output);
 
     assert_eq!(testfile1.contents().as_str(), "AB=DEF\nD=BAR\n\nF=BAR\n");
