+++ tests/fixes/unordered_key.rs
@@ -1,4 +1,4 @@
-use crate::common::TestDir;
+use crate::common::*;
 
 #[test]
 fn unordered_key() {
@@ -27,12 +27,11 @@ fn unordered_key() {
         \n\
         # end comment\n",
     );
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:5 UnorderedKey: The A key should go before the C key\n\
-        .env:13 UnorderedKey: The K key should go before the M key\n\
-        .env:16 UnorderedKey: The I key should go before the K key\n",
-    );
+    let expected_output = fix_output(&[
+        ".env:5 UnorderedKey: The A key should go before the C key",
+        ".env:13 UnorderedKey: The K key should go before the M key",
+        ".env:16 UnorderedKey: The I key should go before the K key",
+    ]);
     testdir.test_command_fix_success(expected_output);
 
     assert_eq!(
